pyproject.toml
[tool.poetry]
name = "sobesity-be"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.2"
pydantic = "^1.10.4"
dependency-injector = "^4.41.0"
sqlalchemy = "2.0.0rc3"
flask-openapi3 = "^2.2.2"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"
pyjwt = "^2.6.0"
bcrypt = "^4.0.1"
alembic = "^1.9.2"
email-validator = "^1.3.1"
http-constants = "^0.5.0"
factory-boy = "^3.2.1"
flask-cors = "^3.0.10"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
ipython = "^8.8.0"
black = "^22.12.0"
flake8 = "^6.0.0"
autoflake = "^2.0.0"
isort = "^5.11.4"
mock = "^5.0.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

